pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["dorvillethiago <dorville.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
psycopg2 = "^2.9.10"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[tool.poetry.scripts]
format = "scripts:format_code"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
